Fix warnings on .NET SDK 6.0 and VS 2022 (#121)
* Fix errors and warnings when building on .NET SDK 6

* Fix indentation snafu in previous commit

* Fix some more warnings in VS

* Fix more analyzer warnings in VS2022

* Work around CA2016 analyzer crashes

* Suppress redundant suppression warnings on .NET SDK 6

It looks like several of the issues worked around for .NET SDK 5.0 are no longer a problem, but that means all the suppressions now get flagged as redundant.

At some point we will require .NET SDK 6, and then we can git rid of these.

* Clean up a few remaining analyzer warnings on VS2022

* Remove private coverlet NuGet feed

This was causing the CI build to fail because the feed was producing 503 errors, and the version we require seems to be available publicly in NuGet.org anyway.

* Make TestReactivePlatformClient honour cleanupEnvironment

This class had been storing the argument but then cleaning up in Dispose in all cases. It now cleans up in Dispose only when asked to. This avoids unnecessary work in tests that create environments that are not shared, and therefore can be discarded without cleanup.

* Discard unused return value

This was causing a code analyser warning. It had been in there to support some debugging scenarios, but it's not required.

* Updated build to handle 'nbgv' already installed

* Make last build fix cross-platform

* switch powershell task

Co-authored-by: James Dawson <james@readsource.co.uk>
Co-authored-by: James Dawson <james.dawson@endjin.com>