Deal with IDE0220 (implicit cast in foreach)
In most places we supress the issue because adding the Cast<T> operator isn't performance neutral: it allocates an additional IEnumerable<T>. The main purpose of the suggestion is to highlight the fact that a cast occurs, and the suppression comments also serve that purpose.

In cases where the performance is definitely not critical (because it's a console app that never gets used as a library) I've gone with the Cast<T> because it is more readable than the suppressions.