Address IDE0251 with readonly annotations
The .NET 7.0.3xx SDK reports IDE0251 (unhelpfully with no error text, but that's a known bug) now for all struct members that don't actually change the struct but are not marked as readonly. (The readonly enables the compiler to avoid defensive copies when using these members.)

This is technically a breaking change on public types, because it is a change in API surface area. As far as I know, though, it is binary compatible. (Unless there's a modopts thing going on to include readonlyness for overload resolution purposes.) It changes the code gen for call sites, but not in any way that should change behaviour. But in any case it's OK because we're still a preview.