Suppress IDE0270
As reported in https://github.com/dotnet/roslyn/issues/68313 there's no way to say that you want this for field initialization but not when checking for nulls for exception throwing purposes.

Maybe we could write a diagnostic suppressor as suggested, but unless/until we do that, this needs to be disabled.