Reinstate dotnet format with workarounds (#140)
* Reinstate dotnet-format but with exclusion

* Fix issues discovered by newer dotnet format.

We recently discovered that the dotnet format build step had been loading an old version of the tool. It had a step to install the tool, but since the tool has been integrated with the .NET SDK since v6.0, the downloadable version was no longer being updated. The upshot is that we were using the .NET 5.0 version of the tool!

The newer tool finds a few additional issues, but also gets confused by the use of aliasing in some files. (We sometimes refer to Object because that has been aliased to ObjectSlim, for example.)

This change suppresses the warning in cases where it arises from Reaqtive's unusual use of aliasing, and fixes a few cases where the newer analyzers had spotted genuine issues.