Upgrade Nerdbank.GitVersioning to 3.6.133
Turns out that the error we got in the VB.NET project when we last tried this was caused by <LangVersion> being set globally to 11.0. That is the latest version of C#, but it's quite an old version of VB.NET, and NBGV turns out to rely on a language feature from a newer version.